Projects/AISoftwareDevTeam/pyproject.toml
[project]
name = "aisoftwaredevteam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"python-dotenv==1.2.1",
	"gradio==4.44.1",
	"huggingface-hub<1.0",
	"openai-agents",
]
